Added animations, components, box2d, and singletons

+++ Engine/Component/PhysicsComponent.cpp
@@ -9,7 +9,7 @@ namespace nh
 	{
 		velocity += acceleration * owner->scene->engine->time.deltaTime;
 		owner->transform.position += velocity * owner->scene->engine->time.deltaTime;
-		owner->transform.position.x = nh::Wrap((int)owner->transform.position.x, -10, 810);
-		owner->transform.position.y = nh::Wrap((int)owner->transform.position.y, -10, 810);
+		owner->transform.position.x = (float)nh::Wrap((int)owner->transform.position.x, -10, 810);
+		owner->transform.position.y = (float)nh::Wrap((int)owner->transform.position.y, -10, 810);
 	}
 }
